LayoutTests/fast/text/font-promises-gc-expected.txt

CONSOLE MESSAGE: line 34: Font status: unloaded
CONSOLE MESSAGE: line 36: Font loaded
This test should not time out, and should say "SUCCESS".
Hello
SUCCESS
